+++ lib/eufemia.dart
@@ -3,10 +3,7 @@ library eufemia;
 // Theme and styles
 export 'src/theme.dart';
 export 'src/style/colors.dart';
-export 'src/utils/icon_position.dart';
-export 'src/utils/button_size.dart';
-export 'src/utils/status.dart';
-
+export 'src/utils/utils.dart';
 // Buttons
 export 'src/buttons/buttons.dart';