+++ lib/src/components/components.dart
@@ -21,3 +21,4 @@ export 'shimmer.dart';
 export 'input_field.dart';
 export 'checkmark.dart';
 export 'segmented_control.dart';
+export 'gradient_widget.dart';