+++ lib/src/components/segmented_control.dart
@@ -17,26 +17,43 @@ class SegmentedControl<T> extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return CustomSlidingSegmentedControl(
-      children: addPadding(children),
-      onValueChanged: onValueChanged,
-      groupValue: groupValue,
-      padding: const EdgeInsets.all(2.0),
-      backgroundColor:
-          context.bright ? EufemiaColors.outlineGray : EufemiaColors.gray4,
-      thumbColor: context.bright ? Colors.white : EufemiaColors.gray6,
-      thumbRadius: context.cupertino ? 6.93 : 100.0,
-      trackRadius: context.cupertino ? 8.0 : 100.0,
+    return DefaultTextStyle(
+      style: Theme.of(context).textTheme.subtitle2.copyWith(
+            color:
+                context.bright ? EufemiaColors.coal : EufemiaColors.lightGray,
+          ),
+      child: CustomSlidingSegmentedControl(
+        children: addPadding(context, children),
+        onValueChanged: onValueChanged,
+        groupValue: groupValue,
+        padding: EdgeInsets.symmetric(
+          vertical: 2.0,
+          horizontal: 3.0,
+        ),
+        backgroundColor:
+            context.bright ? EufemiaColors.outlineGray : EufemiaColors.gray4,
+        thumbColor: context.bright ? Colors.white : EufemiaColors.gray6,
+        thumbRadius: context.cupertino ? 6.93 : 100.0,
+        trackRadius: context.cupertino ? 8.0 : 100.0,
+        border: context.cupertino
+            ? null
+            : Border.all(
+                width: 0.5,
+                color: context.bright
+                    ? EufemiaColors.darkGray
+                    : EufemiaColors.gray2,
+              ),
+      ),
     );
   }
 
-  addPadding(Map<T, Widget> map) {
+  addPadding(BuildContext context, Map<T, Widget> map) {
     return map.map((key, value) => MapEntry(
           key,
           Padding(
-            padding: const EdgeInsets.symmetric(
-              vertical: 6.0,
-              horizontal: 8.0,
+            padding: EdgeInsets.symmetric(
+              vertical: context.cupertino ? 6 : 8.0,
+              horizontal: 32.0 + (context.cupertino ? 8.0 : 16.0),
             ),
             child: value,
           ),