+++ lib/src/graphics/graphics.dart
@@ -9,3 +9,4 @@ export 'cards/card_logo.dart';
 export 'cards/card_preset.dart';
 export 'cards/card_program.dart';
 export 'cards/card_type.dart';
+export 'shadows.dart';